+++ site/pages/index.tsx
@@ -3,6 +3,7 @@ import Head from "next/head";
 import { DrumMachineExample } from "src/DrumMachineExample";
 import { ElectricPianoExample } from "src/ElectricPianoExample";
 import { MalletExample } from "src/MalletExample";
+import { MellotronExample } from "src/MellotronExample";
 import { SoundfontExample } from "src/SoundfontExample";
 import { PianoExample } from "../src/PianoExample";
 
@@ -19,7 +20,7 @@ export default function Home() {
       <main className={"max-w-4xl mx-auto my-20 p-4" + inter.className}>
         <div className="flex items-end mb-16">
           <h1 className="text-6xl font-bold">smplr</h1>
-          <div>0.7.0</div>
+          <div>0.8.0</div>
         </div>
 
         <div className="flex flex-col gap-8">
@@ -28,6 +29,7 @@ export default function Home() {
           <ElectricPianoExample />
           <MalletExample />
           <DrumMachineExample />
+          <MellotronExample />
         </div>
       </main>
     </>