+++ src/sfz/sfz-sampler.ts
@@ -1,5 +1,5 @@
+import { DefaultPlayer } from "../player/default-player";
 import { toMidi } from "../player/midi";
-import { Player } from "../player/player";
 import { SampleStart, SampleStop } from "../player/types";
 import { HttpStorage, Storage } from "../storage";
 import { SfzInstrument } from "./sfz-kits";
@@ -29,21 +29,21 @@ const EMPTY_WEBSFZ: Websfz = Object.freeze({
 
 export class SfzSampler {
   public readonly options: Readonly<Partial<SfzSamplerConfig>>;
-  private readonly player: Player;
+  private readonly player: DefaultPlayer;
   #websfz: Websfz;
-  #load: Promise<void>;
+  public readonly load: Promise<this>;
 
   constructor(
     public readonly context: AudioContext,
     options: Partial<SfzSamplerConfig> & Pick<SfzSamplerConfig, "instrument">
   ) {
     this.options = Object.freeze(Object.assign({}, options));
-    this.player = new Player(context, options);
+    this.player = new DefaultPlayer(context, options);
     this.#websfz = EMPTY_WEBSFZ;
 
     const storage = options.storage ?? HttpStorage;
-    this.#load = loadSfzInstrument(options.instrument, storage).then(
-      (result) => {
+    this.load = loadSfzInstrument(options.instrument, storage)
+      .then((result) => {
         this.#websfz = Object.freeze(result);
         return loadSfzBuffers(
           context,
@@ -51,8 +51,8 @@ export class SfzSampler {
           this.#websfz,
           storage
         );
-      }
-    );
+      })
+      .then(() => this);
   }
 
   get output() {
@@ -60,8 +60,8 @@ export class SfzSampler {
   }
 
   async loaded() {
-    await this.#load;
-    return this;
+    console.warn("deprecated: use load instead");
+    return this.load;
   }
 
   start(sample: SampleStart | string | number) {