+++ src/test/audio-context-test-helpers.ts
@@ -1,14 +1,13 @@
-// Create a mock for the GainNodeMock
-class GainNodeMock {
-  gain: { value: number };
+function add<T>(item: T, list: T[]) {
+  list.push(item);
+  return item;
+}
+
+class NodeMock {
   connected: any[] = [];
   disconnected: any[] = [];
 
-  constructor(public readonly context: AudioContextMock) {
-    this.gain = {
-      value: 1.0,
-    };
-  }
+  constructor(public readonly context: AudioContextMock) {}
 
   connect(destination: any) {
     this.connected.push(destination);
@@ -19,13 +18,53 @@ class GainNodeMock {
   }
 }
 
-function add<T>(item: T, list: T[]) {
-  list.push(item);
-  return item;
+// Create a mock for the GainNodeMock
+class GainNodeMock extends NodeMock {
+  gain: { value: number };
+
+  constructor(context: AudioContextMock) {
+    super(context);
+    this.gain = {
+      value: 1.0,
+    };
+  }
+}
+
+class BufferSourceMock extends NodeMock {
+  detune: {
+    value: number;
+  };
+  startedAt: number | undefined;
+  buffer: AudioBufferMock | undefined;
+
+  constructor(context: AudioContextMock) {
+    super(context);
+    this.detune = {
+      value: 0,
+    };
+  }
+
+  start(time?: number) {
+    if (this.startedAt !== undefined) throw new Error("Already started");
+    this.startedAt = time;
+  }
+}
+
+class AudioBufferMock {
+  constructor(
+    public numberOfChannels: number,
+    public length: number,
+    public sampleRate: number
+  ) {}
+
+  get buffer() {
+    return this as unknown as AudioBuffer;
+  }
 }
 
 class AudioContextMock {
   gains: GainNodeMock[] = [];
+  bufferSources: BufferSourceMock[] = [];
   currentTime = 0;
   destination: any;
 
@@ -35,6 +74,10 @@ class AudioContextMock {
     };
   }
 
+  get context() {
+    return this as unknown as AudioContext;
+  }
+
   createGain() {
     return add(new GainNodeMock(this), this.gains);
   }
@@ -42,11 +85,24 @@ class AudioContextMock {
   decodeAudioData(arrayBuffer: any) {
     return Promise.resolve({ arrayBuffer });
   }
+
+  createBuffer(
+    numberOfChannels: number,
+    length: number,
+    sampleRate: number
+  ): AudioBuffer {
+    return new AudioBufferMock(numberOfChannels, length, sampleRate).buffer;
+  }
+
+  createBufferSource() {
+    return add(new BufferSourceMock(this), this.bufferSources);
+  }
 }
 
 export function createAudioContextMock() {
   (global as any).GainNode = GainNodeMock;
-  return new AudioContextMock() as unknown as AudioContext;
+  (global as any).BufferSourceMock = BufferSourceMock;
+  return new AudioContextMock();
 }
 
 export function createFetchMock(data: Record<string, any>) {